Fixes for windows warnings savannah task #15808
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@38203 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 8773f9e5a9ee90802663d6fe0435722209053d66

+++ Rec/Tr/TrackUtils/src/TrackHitAdder.cpp
@@ -167,8 +167,8 @@ namespace {
 	    typename TYPE::HitRangeType thesehits = hitcreator.hitsLocalXRange( istation, ilayer, iregion, 
 										localx -  tolx,
 										localx +  tolx) ;
-	    Gaudi::XYZPointF trkpoint(state.x(),state.y(),state.z()) ;
-	    Gaudi::XYZVectorF trkdir(state.tx(),state.ty(),1) ;
+	    Gaudi::XYZPointF trkpoint((float) state.x(),(float) state.y(),(float) state.z()) ;
+	    Gaudi::XYZVectorF trkdir((float) state.tx(),(float) state.ty(),1) ;
 	    // if there is more than one hit, choose the closest?
 	    lhcbids.reserve( lhcbids.size() + thesehits.size() ) ;
 	    BOOST_FOREACH( const typename TYPE::HitType* hit, thesehits )
